# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signup": {
      "post": {
        "description": "Always answers the same way, whether or not the address is already registered: a registration form that says 'already taken' is a free account checker. Blocked by the age gate (403 auth/age_restricted) for India under 18 and the EU under 16.",
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 254,
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "locale": {
                    "type": "string",
                    "maxLength": 16
                  },
                  "dateOfBirth": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "jurisdiction": {
                    "type": "string",
                    "enum": [
                      "IN",
                      "EU",
                      "OTHER"
                    ]
                  },
                  "consents": {
                    "type": "object",
                    "properties": {
                      "analytics": {
                        "type": "boolean"
                      },
                      "memory": {
                        "type": "boolean"
                      },
                      "marketing": {
                        "type": "boolean"
                      }
                    }
                  }
                },
                "required": [
                  "email",
                  "password",
                  "dateOfBirth",
                  "jurisdiction"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Verification email sent.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "verification_sent"
                      ]
                    },
                    "email": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "email"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "auth/age_restricted"
          },
          "429": {
            "description": "common/rate_limited; carries Retry-After."
          }
        },
        "summary": "Create an account",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 16,
                    "maxLength": 256
                  }
                },
                "required": [
                  "token"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Confirm an email address with a single-use token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "One message for an unknown address and a wrong password alike, and the same argon2 cost is paid either way.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 254,
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "kind": {
                    "type": "string",
                    "enum": [
                      "web",
                      "desktop",
                      "bridge",
                      "premiere",
                      "ae",
                      "resolve",
                      "api"
                    ]
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access and refresh tokens.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "tokenType": {
                      "type": "string",
                      "enum": [
                        "Bearer"
                      ]
                    },
                    "sessionId": {
                      "type": "string",
                      "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                    },
                    "workspaceId": {
                      "type": "string",
                      "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "owner",
                        "admin",
                        "editor",
                        "viewer"
                      ]
                    },
                    "kind": {
                      "type": "string",
                      "enum": [
                        "web",
                        "desktop",
                        "bridge",
                        "premiere",
                        "ae",
                        "resolve",
                        "api"
                      ]
                    }
                  },
                  "required": [
                    "accessToken",
                    "refreshToken",
                    "expiresIn",
                    "tokenType",
                    "sessionId",
                    "workspaceId",
                    "role",
                    "kind"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "auth/invalid_credentials"
          }
        },
        "summary": "Sign in with email and password",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/magic-link": {
      "post": {
        "operationId": "AuthController_requestMagicLink",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 254,
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Accepted, whether or not the address has an account."
          }
        },
        "summary": "Email a single-use sign-in link (15 minutes)",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/magic-link/consume": {
      "post": {
        "operationId": "AuthController_consumeMagicLink",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 16,
                    "maxLength": 256
                  },
                  "kind": {
                    "type": "string",
                    "enum": [
                      "web",
                      "desktop",
                      "bridge",
                      "premiere",
                      "ae",
                      "resolve",
                      "api"
                    ]
                  }
                },
                "required": [
                  "token"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access and refresh tokens.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "tokenType": {
                      "type": "string",
                      "enum": [
                        "Bearer"
                      ]
                    },
                    "sessionId": {
                      "type": "string",
                      "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                    },
                    "workspaceId": {
                      "type": "string",
                      "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "owner",
                        "admin",
                        "editor",
                        "viewer"
                      ]
                    },
                    "kind": {
                      "type": "string",
                      "enum": [
                        "web",
                        "desktop",
                        "bridge",
                        "premiere",
                        "ae",
                        "resolve",
                        "api"
                      ]
                    }
                  },
                  "required": [
                    "accessToken",
                    "refreshToken",
                    "expiresIn",
                    "tokenType",
                    "sessionId",
                    "workspaceId",
                    "role",
                    "kind"
                  ]
                }
              }
            }
          }
        },
        "summary": "Exchange a magic-link token for tokens",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Rotation with a 60-second grace: presenting the previous token inside the window replays the same new pair, presenting it after the window revokes the whole family.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "minLength": 16,
                    "maxLength": 256
                  }
                },
                "required": [
                  "refreshToken"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A rotated token pair.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "tokenType": {
                      "type": "string",
                      "enum": [
                        "Bearer"
                      ]
                    },
                    "sessionId": {
                      "type": "string",
                      "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                    },
                    "workspaceId": {
                      "type": "string",
                      "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "owner",
                        "admin",
                        "editor",
                        "viewer"
                      ]
                    },
                    "kind": {
                      "type": "string",
                      "enum": [
                        "web",
                        "desktop",
                        "bridge",
                        "premiere",
                        "ae",
                        "resolve",
                        "api"
                      ]
                    }
                  },
                  "required": [
                    "accessToken",
                    "refreshToken",
                    "expiresIn",
                    "tokenType",
                    "sessionId",
                    "workspaceId",
                    "role",
                    "kind"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "common/unauthorized or auth/session_revoked"
          }
        },
        "summary": "Rotate a refresh token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Takes the refresh token rather than the access token, so a client whose access token has already expired can still sign out. Always 204: an unknown token must not be a way to test whether a token is live.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "minLength": 16,
                    "maxLength": 256
                  }
                },
                "required": [
                  "refreshToken"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Signed out."
          }
        },
        "summary": "Revoke a refresh-token family",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/token/exchange": {
      "post": {
        "description": "The workspace is bound into the access token; there is no X-Workspace-Id header (07 section Conventions, THREAT-MODEL T4). Membership is re-checked here.",
        "operationId": "AuthController_exchange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workspaceId": {
                    "type": "string",
                    "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                  }
                },
                "required": [
                  "workspaceId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens for the requested workspace.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "tokenType": {
                      "type": "string",
                      "enum": [
                        "Bearer"
                      ]
                    },
                    "sessionId": {
                      "type": "string",
                      "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                    },
                    "workspaceId": {
                      "type": "string",
                      "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "owner",
                        "admin",
                        "editor",
                        "viewer"
                      ]
                    },
                    "kind": {
                      "type": "string",
                      "enum": [
                        "web",
                        "desktop",
                        "bridge",
                        "premiere",
                        "ae",
                        "resolve",
                        "api"
                      ]
                    }
                  },
                  "required": [
                    "accessToken",
                    "refreshToken",
                    "expiresIn",
                    "tokenType",
                    "sessionId",
                    "workspaceId",
                    "role",
                    "kind"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "auth/not_a_member"
          }
        },
        "summary": "Switch the active workspace",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/sessions/{sessionId}": {
      "delete": {
        "operationId": "AuthController_revokeSession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked."
          }
        },
        "summary": "Revoke one session (and its refresh-token family)",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/parental-waitlist": {
      "post": {
        "description": "Offered to a sign-up blocked by the age gate (D60), until a verifiable parental-consent flow ships.",
        "operationId": "AuthController_joinWaitlist",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 254,
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "summary": "Join the parental-consent waitlist",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Aksharo API",
    "description": "Errors use the envelope `{ error: { code, message, details?, requestId } }` with `namespace/slug` codes (CONTRACTS section 8).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    },
...
  }
}